Moved the gathering of the options in the experiment file;
